Responded to PR comments
